TestFulProof/requirements.txt
pip
selenium
pytest
allure-pytest